Create responsivewebsite.html
A piece of code in the HTML body element in the below Code Playground is the Bootstrap Navbar without a list-based approach Code.
The Bootstrap class name container provides us default left and right spacings starting from smaller devices for a better user experience. It has one fixed width for each breakpoint in Bootstrap (extra small, small, medium, large, and extra large).
The Bootstrap class name container-fluid is a full width container, spanning the entire width of the viewport.

A piece of the code in the HTML body element is the Bootstrap arrow-right-short icon code in the below Code Playground.